+++ test/connectionmonitor/SignalingAndMetricsConnectionMonitor.test.ts
@@ -124,6 +124,8 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     videoDownstreamFrameWidth: RawMetrics = 100;
     audioPacketsSent: RawMetrics = 50;
     totalBytesReceived: number = 0;
+    totalRequestsReceived: number = 0;
+    totalResponsesReceived: number = 0;
 
     getObservableMetrics(): { [id: string]: number } {
       return {
@@ -155,6 +157,8 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
             type: 'candidate-pair',
             ...{
               bytesReceived: this.totalBytesReceived,
+              requestsReceived: this.totalRequestsReceived,
+              responsesReceived: this.totalResponsesReceived,
             },
           },
         ],
@@ -164,6 +168,8 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
             type: 'candidate-pair',
             ...{
               bytesReceived: 0,
+              requestsReceived: 0,
+              responsesReceived: 0,
             },
           },
         ],
@@ -319,7 +325,7 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     expect(lastPacketLossInboundTimestampMsCalled).to.equal(false);
   });
 
-  it('can return without changing stats when total bytes received is negative', () => {
+  it('can return without changing stats when total packets received is negative', () => {
     testClientMetricReport.totalBytesReceived = 4;
     testClientMetricReport.fractionLoss = 0;
     testClientMetricReport.audioPacketsReceived = 1;
@@ -333,7 +339,7 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     expect(lastPacketLossInboundTimestampMsCalled).to.equal(false);
   });
 
-  it('can reset and increment consecutive stats with no bytes when bytes received are followed by no bytes', () => {
+  it('can reset and increment consecutive stats with no packets when packets received are followed by no packets', () => {
     testClientMetricReport.totalBytesReceived = 1;
     testClientMetricReport.fractionLoss = 0;
     testClientMetricReport.audioPacketsReceived = 1;
@@ -341,14 +347,45 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     expect(consecutiveStatsWithNoPackets).to.equal(0);
     testClientMetricReport.totalBytesReceived = 1;
     testClientMetricReport.fractionLoss = 0;
-    testClientMetricReport.audioPacketsReceived = 0;
+    testClientMetricReport.audioPacketsReceived = 1;
     sendClientMetricReport(testClientMetricReport);
     expect(consecutiveStatsWithNoPackets).to.equal(1);
     testClientMetricReport.totalBytesReceived = 1;
     testClientMetricReport.fractionLoss = 0;
+    testClientMetricReport.audioPacketsReceived = 1;
+    sendClientMetricReport(testClientMetricReport);
+    expect(consecutiveStatsWithNoPackets).to.equal(2);
+  });
+
+  it('can reset and increment consecutive stats with no packets when stun packets received are followed by no packets', () => {
+    testClientMetricReport.totalBytesReceived = 0;
+    testClientMetricReport.totalRequestsReceived = 1;
+    testClientMetricReport.totalResponsesReceived = 0;
+    testClientMetricReport.fractionLoss = 0;
+    testClientMetricReport.audioPacketsReceived = 0;
+    sendClientMetricReport(testClientMetricReport);
+    expect(consecutiveStatsWithNoPackets).to.equal(0);
+    testClientMetricReport.totalBytesReceived = 0;
+    testClientMetricReport.totalRequestsReceived = 1;
+    testClientMetricReport.totalResponsesReceived = 0;
+    testClientMetricReport.fractionLoss = 0;
+    testClientMetricReport.audioPacketsReceived = 0;
+    sendClientMetricReport(testClientMetricReport);
+    expect(consecutiveStatsWithNoPackets).to.equal(1);
+    testClientMetricReport.totalBytesReceived = 0;
+    testClientMetricReport.totalRequestsReceived = 1;
+    testClientMetricReport.totalResponsesReceived = 0;
+    testClientMetricReport.fractionLoss = 0;
     testClientMetricReport.audioPacketsReceived = 0;
     sendClientMetricReport(testClientMetricReport);
     expect(consecutiveStatsWithNoPackets).to.equal(2);
+    testClientMetricReport.totalBytesReceived = 0;
+    testClientMetricReport.totalRequestsReceived = 1;
+    testClientMetricReport.totalResponsesReceived = 1;
+    testClientMetricReport.fractionLoss = 0;
+    testClientMetricReport.audioPacketsReceived = 0;
+    sendClientMetricReport(testClientMetricReport);
+    expect(consecutiveStatsWithNoPackets).to.equal(0);
   });
 
   it('can set last packet loss inbound timestamp due to no packets', () => {